refactor: Rename tooltip and make generic
